refactor container image code and use in gradio, streamlit, dash and shiny apps
