+++ apps/forms/dash.py
@@ -7,16 +7,16 @@ from django.utils.safestring import mark_safe
 
 from apps.forms.base import AppBaseForm
 from apps.forms.field.common import SRVCommonDivField
+from apps.forms.mixins import ContainerImageMixin
 from apps.models import DashInstance
 from projects.models import Flavor
 
 __all__ = ["DashForm"]
 
 
-class DashForm(AppBaseForm):
+class DashForm(ContainerImageMixin, AppBaseForm):
     flavor = forms.ModelChoiceField(queryset=Flavor.objects.none(), required=False, empty_label=None)
     port = forms.IntegerField(min_value=3000, max_value=9999, required=True)
-    image = forms.CharField(max_length=255, required=True)
     default_url_subpath = forms.CharField(max_length=255, required=False, label="Custom URL subpath")
 
     def _setup_form_fields(self):
@@ -32,6 +32,9 @@ class DashForm(AppBaseForm):
                 f" on the Serve <a href='{apps_url}'>Apps & Models</a> page."
             )
         )
+
+        # Setup container image field from mixin
+        self._setup_container_image_field()
 
     def _setup_form_helper(self):
         super()._setup_form_helper()
@@ -50,7 +53,8 @@ class DashForm(AppBaseForm):
             ),
             SRVCommonDivField("source_code_url", placeholder="Provide a link to the public source code"),
             SRVCommonDivField("port", placeholder="8000"),
-            SRVCommonDivField("image", placeholder="e.g. docker.io/username/image-name:image-tag"),
+            # Container image field
+            self._setup_container_image_helper(),
             Accordion(
                 AccordionGroup(
                     "Advanced settings",
